pyproject.toml
[tool.poetry]
name = "helloyaponiya"
version = "0.1.0"
description = ""
authors = ["izumo-hackathon-01-team-C"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
pydantic = "^2.6.2"
uvicorn = "^0.27.1"
tortoise-orm = "^0.20.0"
pysqlite3 = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
